Workaround for libhoudini.so sin&cos bug
This is a workaround for Intel ARM->x86 translator (houdini) bug.
sin() & cos() return garbage for very large inputs. The outcome is
that when codomain is applied the result interval becomes empty.

Workaround is to bring the input value to the base range via modulo
if sin/cos returns an invalid values.

Change-Id: I2178daf2f9ffec543c1cf9fad2881719dd7736cc